refactor: moves retry helper to subfolder
